custom adapter for json decimal types and add mongodb testcontainer

+++ app/src/main/kotlin/botlinera/domain/valueobject/Coordinates.kt
@@ -2,6 +2,6 @@ package botlinera.domain.valueobject
 
 import java.math.BigDecimal
 
-class Coordinates(latitude: BigDecimal, longitude: BigDecimal) {
+class Coordinates(latitude: Double, longitude: Double) {
 
 }
